Cache to conftest
